Field Validation Completed

+++ src/main/java/com/springboot/bookmyshow/service/MovieService.java
@@ -5,8 +5,12 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.springboot.bookmyshow.dao.AdminDao;
 import com.springboot.bookmyshow.dao.MovieDao;
+import com.springboot.bookmyshow.entity.Admin;
 import com.springboot.bookmyshow.entity.Movie;
+import com.springboot.bookmyshow.exception.LoginFailed;
+import com.springboot.bookmyshow.exception.MovieNotFound;
 import com.springboot.bookmyshow.util.ResponseStructure;
 
 @Service
@@ -15,13 +19,20 @@ public class MovieService
 	@Autowired
 	MovieDao movieDao;
 	
-	public ResponseEntity<ResponseStructure<Movie>> saveMovie(Movie movie)
+	@Autowired
+	AdminDao adminDao;
+	
+	public ResponseEntity<ResponseStructure<Movie>> saveMovie(String adminEmail,String adminPassword,Movie movie)
 	{
-		ResponseStructure<Movie> structure = new ResponseStructure<Movie>();
-		structure.setMessage("Movie was Saved...");
-		structure.setStatus(HttpStatus.CREATED.value());
-		structure.setData(movieDao.saveMovie(movie));
-		return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.CREATED);
+		Admin exAdmin = adminDao.adminLogin(adminEmail, adminPassword);
+		if(exAdmin!=null) {
+			ResponseStructure<Movie> structure = new ResponseStructure<Movie>();
+			structure.setMessage("Movie was Saved...");
+			structure.setStatus(HttpStatus.CREATED.value());
+			structure.setData(movieDao.saveMovie(movie));
+			return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.CREATED);
+		}
+		throw new LoginFailed("Admin Login Failed.....");
 	}
 	
 	public ResponseEntity<ResponseStructure<Movie>> findMovie(int movieId)
@@ -35,34 +46,42 @@ public class MovieService
 			structure.setData(movie);
 			return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.FOUND);
 		}
-		return null;
+		throw new LoginFailed("Admin Login Failed.....");
 	}
 	
-	public ResponseEntity<ResponseStructure<Movie>> deleteMovie(int movieId)
+	public ResponseEntity<ResponseStructure<Movie>> deleteMovie(String adminEmail,String adminPassword,int movieId)
 	{
-		ResponseStructure<Movie> structure = new ResponseStructure<Movie>();
-		Movie movie = movieDao.deleteMovie(movieId);
-		if(movie!=null) {
-			structure.setMessage("Data Deleted...");
-			structure.setStatus(HttpStatus.OK.value());
-			structure.setData(movie);
-			return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.OK);
+		Admin exAdmin = adminDao.adminLogin(adminEmail, adminPassword);
+		if(exAdmin!=null) {
+			ResponseStructure<Movie> structure = new ResponseStructure<Movie>();
+			Movie movie = movieDao.deleteMovie(movieId);
+			if(movie!=null) {
+				structure.setMessage("Data Deleted...");
+				structure.setStatus(HttpStatus.OK.value());
+				structure.setData(movie);
+				return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.OK);
+			}
+			throw new MovieNotFound("Movie not Found with the given id...");
 		}
-		return null;
+		throw new LoginFailed("Admin Login Failed.....");
 	}
 	
-	public ResponseEntity<ResponseStructure<Movie>> updateMovie(Movie movie, int movieId)
+	public ResponseEntity<ResponseStructure<Movie>> updateMovie(String adminEmail,String adminPassword,Movie movie, int movieId)
 	{
-		ResponseStructure<Movie> structure = new ResponseStructure<Movie>();
-		Movie exMovie = movieDao.updateMovie(movie, movieId);
-		if(exMovie!=null)
-		{
-			structure.setMessage("Data Updated...");
-			structure.setStatus(HttpStatus.OK.value());
-			structure.setData(exMovie);
-			return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.OK);
+		Admin exAdmin = adminDao.adminLogin(adminEmail, adminPassword);
+		if(exAdmin!=null) {
+			ResponseStructure<Movie> structure = new ResponseStructure<Movie>();
+			Movie exMovie = movieDao.updateMovie(movie, movieId);
+			if(exMovie!=null)
+			{
+				structure.setMessage("Data Updated...");
+				structure.setStatus(HttpStatus.OK.value());
+				structure.setData(exMovie);
+				return new ResponseEntity<ResponseStructure<Movie>>(structure,HttpStatus.OK);
+			}
+			throw new MovieNotFound("Movie not Found with the given id...");
 		}
-		return null;
+		throw new LoginFailed("Admin Login Failed.....");
 	}
 
 }
